backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=11.0.0
rembg>=2.0.68
vtracer>=0.6.0
scour>=0.38.2
httpx>=0.28.0
pydantic>=2.5.0
